+++ section/section-featured.php
@@ -3,7 +3,7 @@
 		<div class="<?=$rowClass?>"> 
 			<div class="span12">  
 	            <h1>Simple <span class="lighter">&amp; Intuitive</span></h1>
-	            <p class="light"><em>Minimalistic Website Theme Created for<br>Creative individuals and Startups</em></p>
+	            <p class="light">Minimalistic Website Theme Created for<br>Creative individuals and Startups</p>
 	            <p><a class="btn btn-info btn-large fancybox video" href="<?php echo FEATURE_VIDEO_URL; ?>">Quick Tour &raquo;</a></p>
 	        </div>    
         </div>